refact: use Robolectric at :platform:navigation-test
